move UserInfo, UserDetails to a common place in models
