+++ src/app/containers/app/app.component.ts
@@ -80,8 +80,6 @@ export class AppComponent implements OnInit {
   public userDetailsHandler(userDetails) {
 
     if (propsExist(userDetails, ['sessionTimeout'] ) && userDetails.sessionTimeout.totalIdleTime > 0) {
-
-      console.log(userDetails.sessionTimeout);
       const { idleModalDisplayTime, totalIdleTime } = userDetails.sessionTimeout;
 
       this.addTimeoutNotificationServiceListener();
