+++ src/app/services/logger/logger.service.ts
@@ -4,6 +4,7 @@ import { Injectable } from '@angular/core';
 import { environment as config } from '../../../environments/environment';
 import { CryptoWrapper } from './cryptoWrapper';
 import { SessionStorageService } from '../session-storage/session-storage.service';
+import { UserInfo } from '../../models/user-details.model'; 
 
 export interface ILoggerService {
     trace(message: any, ...additional: any[]): void;
@@ -70,7 +71,7 @@ export class LoggerService implements ILoggerService {
     getMessage(message: any): string {
         const userInfoStr = this.sessionStorageService.getItem('userDetails');
         if(userInfoStr) {
-            const userInfo = JSON.parse(userInfoStr);
+            const userInfo: UserInfo = JSON.parse(userInfoStr);
             if(userInfo && userInfo.email) {
                 const userIdEncrypted = this.cryptoWrapper.encrypt(userInfo.email);
                 const encryptedMessage = `User - ${userIdEncrypted.toString()}, Message - ${message}, Timestamp - ${Date.now()}`
