+++ src/app/services/user/user.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { UserDetails } from '../../../app/store/reducers/app-config.reducer';
+import { UserDetails } from '../../../app/models/user-details.model';
 
 // TODO: Write unit tests
 @Injectable({
